refactor(server): delete shared/http-helpers after migration
